v2_services/agent_17_error_handler/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pyyaml>=6.0
pytest>=7.4.0
httpx>=0.26.0
